# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sensible",
        "title": "Sensible Query Spec",
        "contact": {
            "name": "sensiblequery",
            "url": "https://github.com/sensible-contract/sensiblequery",
            "email": "jiedohh@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/ft/balance/{codehash}/{genesis}/{address}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token FT"
                ],
                "summary": "查询某人持有的某FT Token的余额，同时返回UTXO数量",
                "parameters": [
                    {
                        "type": "string",
                        "default": "844c56bb99afc374967a27ce3b46244e2e1fba60",
                        "description": "Code Hash160",
                        "name": "codehash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "74967a27ce3b46244e2e1fba60844c56bb99afc3",
                        "description": "Genesis ID",
                        "name": "genesis",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "17SkEw2md5avVNyYgj6RiXuQKNwkXaxFyQ",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.FTOwnerBalanceWithUtxoCountResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ft/codehash-info/{codehash}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token FT"
                ],
                "summary": "查询使用某codehash的FT Token简述",
                "parameters": [
                    {
                        "type": "string",
                        "default": "844c56bb99afc374967a27ce3b46244e2e1fba60",
                        "description": "Code Hash160",
                        "name": "codehash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.FTInfoResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ft/codehash/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token FT"
                ],
                "summary": "查询所有FT codehash简述",
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.TokenCodeHashResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ft/genesis-info/{codehash}/{genesis}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token FT"
                ],
                "summary": "查询使用某codehash+genesis的FT Token简述",
                "parameters": [
                    {
                        "type": "string",
                        "default": "844c56bb99afc374967a27ce3b46244e2e1fba60",
                        "description": "Code Hash160",
                        "name": "codehash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "74967a27ce3b46244e2e1fba60844c56bb99afc3",
                        "description": "Genesis ID",
                        "name": "genesis",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": {}, \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.FTInfoResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/ft/info/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token FT"
                ],
                "summary": "查询所有FT Token简述",
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.FTInfoResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ft/owners/{codehash}/{genesis}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token FT"
                ],
                "summary": "查询FT Token的持有人。获得每个地址的token余额",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始游标",
                        "name": "cursor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "返回记录数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "844c56bb99afc374967a27ce3b46244e2e1fba60",
                        "description": "Code Hash160",
                        "name": "codehash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "74967a27ce3b46244e2e1fba60844c56bb99afc3",
                        "description": "Genesis ID",
                        "name": "genesis",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.FTOwnerBalanceResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ft/summary-data/{address}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token FT"
                ],
                "summary": "查询某人持有的FT Token列表。获得每个token的余额。并返回token总量",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始游标",
                        "name": "cursor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "返回记录数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "17SkEw2md5avVNyYgj6RiXuQKNwkXaxFyQ",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": {}, \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.FTSummaryDataByAddressResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ft/summary/{address}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token FT"
                ],
                "summary": "查询某人持有的FT Token列表。获得每个token的余额",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始游标",
                        "name": "cursor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "返回记录数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "17SkEw2md5avVNyYgj6RiXuQKNwkXaxFyQ",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.FTSummaryByAddressResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ft/transfer-times/{codehash}/{genesis}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token FT"
                ],
                "summary": "查询FT Token在区块中的转移数量，以合约CodeHash+GenesisID来确认一种FT",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Start Block Height",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "End Block Height",
                        "name": "end",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "844c56bb99afc374967a27ce3b46244e2e1fba60",
                        "description": "Code Hash160",
                        "name": "codehash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "74967a27ce3b46244e2e1fba60844c56bb99afc3",
                        "description": "Genesis ID",
                        "name": "genesis",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.BlockTokenVolumeResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.FTOwnerBalanceResp": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "token持有人的address",
                    "type": "string"
                },
                "balance": {
                    "description": "余额",
                    "type": "integer"
                },
                "decimal": {
                    "description": "decimal",
                    "type": "integer"
                },
                "pendingBalance": {
                    "description": "待确认余额",
                    "type": "integer"
                }
            }
        },
        "model.FTOwnerBalanceWithUtxoCountResp": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "token持有人的address",
                    "type": "string"
                },
                "balance": {
                    "description": "余额",
                    "type": "integer"
                },
                "decimal": {
                    "description": "decimal",
                    "type": "integer"
                },
                "pendingBalance": {
                    "description": "待确认余额",
                    "type": "integer"
                },
                "utxoCount": {
                    "description": "UTXO 数量",
                    "type": "integer"
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.TokenCodeHashResp": {
            "type": "object",
            "properties": {
                "codehash": {
                    "description": "FT合约hash160(CodePart)",
                    "type": "string"
                },
                "count": {
                    "description": "采用当前合约的Token种类数，NFT包括具体NFT数量；FT只包括种类数量",
                    "type": "integer"
                },
                "inTimes": {
                    "description": "总输入次数",
                    "type": "integer"
                },
                "outTimes": {
                    "description": "总输出次数",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}